chore: input의 작업들을 커스텀 훅으로 분리
